fix(modal unit test): file organization
create a spec file for each component and move the unit test to the corresponding file

"fix 1615"
